+++ activesupport/lib/active_support/notifications/instrumenter.rb
@@ -82,11 +82,6 @@ module ActiveSupport
         @allocation_count_finish = now_allocations
       end
 
-      def end=(ending)
-        ActiveSupport::Deprecation.deprecation_warning(:end=, :finish!)
-        @end = ending
-      end
-
       # Returns the CPU time (in milliseconds) passed since the call to
       # +start!+ and the call to +finish!+
       def cpu_time